# openapi.yaml
openapi: 3.1.0
info:
  title: PEG HR LMS v1 API
  version: "1.0.0"
  description: >
    HTTP-контракт системы онбординга новых сотрудников (см. SPEC.md).
    Один сервис. Все пути под /api/v1. Реализовано в prototype/.
    Ошибка любого запроса: { error: { code, message, details? }, request_id }.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Err:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "snake_case машинный код" }
            message: { type: string }
            details: {}
    Stage:
      type: string
      enum: [intern, onboarding, completed, archived]
    ContentType:
      type: string
      enum: [video, pdf, text]
    LessonStatus:
      type: string
      enum: [locked, available, passed]
    Answer:
      type: object
      required: [question_id, option_index]
      properties:
        question_id: { type: string, format: uuid }
        option_index: { type: integer }
    TestResult:
      type: object
      properties:
        score_pct: { type: integer }
        passed: { type: boolean }
        pass_mark_pct: { type: integer }
        total: { type: integer }
        correct: { type: integer }
        attempt_no: { type: integer }
...
    QuestionEmployee:
      description: вопрос без верного ответа (для сотрудника)
      type: object
      properties:
        id: { type: string, format: uuid }
        text: { type: string }
        options: { type: array, items: { type: string } }
...
security:
  - bearer: []
paths:
...
  /me/pre-onboarding/{itemId}/view:
    post:
      tags: [me]
      summary: Отметить материал просмотренным
      parameters: [{ name: itemId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  onboarding_opened: { type: boolean }
  /me/trajectory:
    get:
      tags: [me]
      summary: Траектория сотрудника с прогрессом
      description: 409, если stage = intern (онбординг не открыт).
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  stage: { $ref: '#/components/schemas/Stage' }
                  onboarding_due_date: { type: [string, "null"] }
                  overdue: { type: boolean }
                  progress: { type: object, properties: { passed: { type: integer }, total: { type: integer } } }
                  trajectory:
                    type: [object, "null"]
                    properties:
                      blocks:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            title: { type: string }
                            lessons:
                              type: array
                              items:
                                type: object
                                properties:
                                  id: { type: string }
                                  title: { type: string }
                                  status: { $ref: '#/components/schemas/LessonStatus' }
                                  material_done: { type: boolean }
                                  material_type: { type: [string, "null"] }
                      attestation:
                        type: object
                        properties:
                          available: { type: boolean }
                          passed: { type: boolean }
                          question_count: { type: integer }
                          pass_mark_pct: { type: integer }
        "409": { $ref: '#/components/responses/Err' }
  /me/lessons/{lessonId}:
    get:
      tags: [me]
      summary: Детали урока (материал + тест без верных ответов)
      description: 403, если урок locked.
      parameters: [{ name: lessonId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  title: { type: string }
                  status: { $ref: '#/components/schemas/LessonStatus' }
                  material_done: { type: boolean }
                  material:
                    type: [object, "null"]
                    properties:
                      content_type: { $ref: '#/components/schemas/ContentType' }
                      file_url: { type: [string, "null"] }
                      text_body: { type: [string, "null"] }
                      min_watch_pct: { type: [integer, "null"] }
                  test:
                    type: [object, "null"]
                    properties:
                      id: { type: string }
                      pass_mark_pct: { type: integer }
                      questions: { type: array, items: { $ref: '#/components/schemas/QuestionEmployee' } }
                      last_attempt:
                        type: [object, "null"]
                        properties:
                          attempt_no: { type: integer }
                          score_pct: { type: integer }
                          passed: { type: boolean }
        "403": { $ref: '#/components/responses/Err' }
  /me/lessons/{lessonId}/video-progress:
    post:
      tags: [me]
      summary: Отчёт о просмотре видео (прогресс копится на сервере)
      parameters: [{ name: lessonId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [pct], properties: { pct: { type: integer, minimum: 0, maximum: 100 } } } } } }
      responses:
        "200": { description: "{ video_pct }" }
        "403": { $ref: '#/components/responses/Err' }
  /me/lessons/{lessonId}/material-done:
    post:
      tags: [me]
      summary: Отметить материал урока пройденным (идемпотентно)
      description: >
        Для видео сервер проверяет накопленный video_pct >= min_watch_pct,
        иначе 422 video_not_watched.
      parameters: [{ name: lessonId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ ok: true }" }
        "403": { $ref: '#/components/responses/Err' }
        "422": { $ref: '#/components/responses/Err', description: video_not_watched }
  /me/lessons/{lessonId}/test:
    post:
      tags: [me]
      summary: Сдать тест урока
      parameters: [{ name: lessonId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers: { type: array, items: { $ref: '#/components/schemas/Answer' } }
      responses:
        "200":
          content: { application/json: { schema: { $ref: '#/components/schemas/TestResult' } } }
  /me/attestation:
    get:
      tags: [me]
      summary: Финальный тест аттестации
      description: 403, если не все уроки пройдены.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_id: { type: string }
                  pass_mark_pct: { type: integer }
                  questions: { type: array, items: { $ref: '#/components/schemas/QuestionEmployee' } }
                  last_attempt: { type: [object, "null"] }
                  completed: { type: boolean }
        "403": { $ref: '#/components/responses/Err' }
    post:
      tags: [me]
      summary: Сдать аттестацию (при passed → stage = completed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers: { type: array, items: { $ref: '#/components/schemas/Answer' } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/TestResult' } } } }
        "422": { $ref: '#/components/responses/Err' }

  # ---------------- employees (HR) ----------------
...